Move SettingsResetter to separate module
