+++ pymc3/step_methods/metropolis.py
@@ -12,9 +12,7 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import copy
-
-from typing import Any, List
+from typing import List
 
 import numpy as np
 import numpy.random as nr
@@ -31,6 +29,7 @@ from pymc3.step_methods.arraystep import (
     PopulationArrayStepShared,
     metrop_select,
 )
+from pymc3.step_methods.settings_resetter import SettingsResetter
 from pymc3.theanof import floatX
 
 __all__ = [
@@ -96,28 +95,6 @@ class MultivariateNormalProposal(Proposal):
         else:
             b = np.random.randn(self.n)
             return np.dot(self.chol, b)
-
-
-class SettingNotFoundInAttribute(BaseException):
-    pass
-
-
-class SettingsResetter:
-    """Stores a copy of initial settings so they can be reset on call"""
-
-    initial_settings: Dict[str, Any]
-
-    def __init__(self, step_method: Any, *settings: str):
-        try:
-            self.initial_settings = {
-                param: copy.deepcopy(getattr(step_method, param)) for param in settings
-            }
-        except AttributeError:
-            raise SettingNotFoundInAttribute("check arguments for typos")
-
-    def __call__(self, step_method: Any) -> Any:
-        for param, initial_value in self.initial_settings.items():
-            setattr(step_method, param, copy.deepcopy(initial_value))
 
 
 class Metropolis(ArrayStepShared):
